threatsimai-swarm/requirements.txt
click>=8.1
pyyaml>=6.0
scapy>=2.5
numpy>=1.24
scikit-learn>=1.3
pandas>=2.0
tqdm>=4.66